Move actions on todos into separate custom hook and add style to list of todos
