+++ src/components/TodoApp.js
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
 import { Grid, Paper } from '@material-ui/core';
 import TodoList from './TodoList';
+import useTodoState from '../hooks/useTodoState';
 
 function TodoApp() {
   const initialTodos = [
@@ -8,31 +9,7 @@ function TodoApp() {
     { id: 2, task: 'Meditate', completed: true },
     { id: 3, task: 'Build first project', completed: false }
   ]
-  const [todos, setTodos] = useState(initialTodos);
-  const addTodo = newTask => {
-    if (newTask === "") {
-      return;
-    }
-    const newId = todos.length ? todos[todos.length - 1].id + 1 : 1;
-    const newTodo = { id: newId, task: newTask, completed: false };
-    setTodos([...todos, newTodo]);
-  };
-  const toggleTodo = todoId => {
-    const updatedTodos = todos.map(todo =>
-      todo.id === todoId ? {...todo, completed: !todo.completed} : todo
-    );
-    setTodos(updatedTodos);
-  };
-  const deleteTodo = todoId => {
-    const filteredTodos = todos.filter(todo => todo.id !== todoId);
-    setTodos(filteredTodos);
-  };
-  const editTodo = (todoId, newTask) => {
-    const updatedTodos = todos.map(todo =>
-      todo.id === todoId ? {...todo, task: newTask} : todo
-    );
-    setTodos(updatedTodos);
-  };
+  const {todos, addTodo, toggleTodo, deleteTodo, editTodo} = useTodoState(initialTodos);
   return (
     <Paper elevation={0}>
       <Grid container justifyContent='center'>
